Creates InfoTooltip component
